pyproject.toml
[tool.poetry]
name = "family-tree"
version = "0.1.0"
description = ""
authors = ["AdamKhagar <adamhagarov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-environ = "^0.11.2"
factory-boy = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
